+++ examples/Databases/list_databases.js
@@ -1,0 +1,13 @@
+const scalingo = require('../../dist/scalingo.js')
+
+scalingo
+  .clientFromToken(process.env.SCALINGO_TOKEN)
+  .then((client) => {
+    return client.Databases.all()
+  })
+  .then((databases) => {
+    console.log(databases)
+  })
+  .catch((err) => {
+    console.error(err)
+  })